requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
deep-translator==1.11.4
executing==2.2.1
filelock==3.24.3
fire==0.7.1
fonttools==4.61.1
fsspec==2026.2.0
fugashi==1.5.2
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.3.1
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
idna==3.11
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jaconv==0.5.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.8.0
jupyter_core==5.9.1
kaggle==2.0.0
kagglesdk==0.1.15
kiwisolver==1.4.9
lightning-utilities==0.15.3
loguru==0.7.3
manga-ocr==0.1.14
manga109api==0.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.2
opencv-python-headless==4.13.0.92
packaging==26.0
parso==0.8.6
pillow==12.1.1
platformdirs==4.9.2
prompt_toolkit==3.0.52
protobuf==6.33.5
psutil==7.2.2
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.3
pyzmq==27.1.0
regex==2026.2.19
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scipy==1.17.1
sentry-sdk==2.53.0
setuptools==82.0.0
shellingham==1.5.4
simsimd==6.5.13
six==1.17.0
smmap==5.0.2
soupsieve==2.8.3
stack-data==0.6.3
stringzilla==4.6.0
sympy==1.14.0
termcolor==3.3.0
text-unidecode==1.3
tokenizers==0.22.2
torch==2.10.0
torchmetrics==1.8.2
torchvision==0.25.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
unidic-lite==1.0.8
urllib3==2.6.3
wandb==0.25.0
wcwidth==0.6.0
webencodings==0.5.1
win32_setctime==1.2.0